+++ gherkin/python/test/gherkin_in_markdown_token_matcher_test.py
@@ -2,7 +2,7 @@
 
 import pytest
 from gherkin.token import Token
-from gherkin.token_matcher import GherkinInMarkdownTokenMatcher
+from gherkin.token_matcher_markdown import GherkinInMarkdownTokenMatcher
 from gherkin.gherkin_line import GherkinLine
 location = { 'line': 1, 'column': 1 }
